refactor: error message를 enum으로 도메인 별 관리하도록 수정
