+++ src/main/java/com/thekey/stylekeyserver/auth/service/AuthService.java
@@ -4,6 +4,7 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.thekey.stylekeyserver.auth.AuthErrorMessage;
 import com.thekey.stylekeyserver.auth.entity.AuthPasswordEntity;
 import com.thekey.stylekeyserver.auth.repository.AuthPasswordRepository;
 import com.thekey.stylekeyserver.auth.service.dto.RegisterPasswordAuthRequest;
@@ -25,7 +26,7 @@ public class AuthService {
         AuthPasswordEntity authEntity = authPasswordJpaRepository.findById(request.getId()).orElse(null);
 
         if (authEntity != null) {
-            throw new DataIntegrityViolationException("이미 가입되었습니다.");
+            throw new DataIntegrityViolationException(AuthErrorMessage.ALREADY_REGISTERED.get());
         }
 
         String encryptedPassword = passwordEncoder.encode(request.getPassword());
